+++ src/full/Agda/TypeChecking/Coverage.hs
@@ -939,15 +939,17 @@ data CosplitError
       -- ^ We do not know the target type of the clause.
   | CosplitNoRecordType Telescope Type
       -- ^ Type living in the given telescope is not a record type.
-  | CosplitIrrelevantProjections
-      -- ^ Record has irrelevant fields, but we do not have irrelevant projections.
+  -- Andreas, 2018-06-09, issue #2170: splitting with irrelevant fields is always fine!
+  -- -- | CosplitIrrelevantProjections
+  -- --     -- ^ Record has irrelevant fields, but we do not have irrelevant projections.
 
 instance PrettyTCM CosplitError where
   prettyTCM = \case
     CosplitNoTarget ->
       text "target type is unknown"
-    CosplitIrrelevantProjections ->
-      text "record has irrelevant fields, but no corresponding projections"
+    -- Andreas, 2018-06-09, issue #2170: splitting with irrelevant fields is always fine!
+    -- CosplitIrrelevantProjections ->
+    --   text "record has irrelevant fields, but no corresponding projections"
     CosplitNoRecordType tel t -> addContext tel $ do
       text "target type " <+> prettyTCM t <+> text " is not a record type"
 
@@ -975,9 +977,11 @@ splitResult f sc@(SClause tel ps _ _ target) = do
           , text   "applied to parameters vs = " <+> (addContext tel $ prettyTCM vs)
           , text $ "and have fields       fs = " ++ prettyShow fs
           ]
-        -- Andreas, 2018-03-19, issue #2971, check that we have a "strong" record type,
-        -- i.e., with all the projections.  Otherwise, we may not split.
-        ifNotM (strongRecord fs) (failure CosplitIrrelevantProjections) $ {-else-} do
+        -- Andreas, 2018-06-09, issue #2170, we always have irrelevant projections
+        -- available on the lhs.
+        -- -- Andreas, 2018-03-19, issue #2971, check that we have a "strong" record type,
+        -- -- i.e., with all the projections.  Otherwise, we may not split.
+        -- ifNotM (strongRecord fs) (failure CosplitIrrelevantProjections) $ {-else-} do
         let es = patternsToElims $ fromSplitPatterns ps
         -- Note: module parameters are part of ps
         let self  = defaultArg $ Def f [] `applyE` es
@@ -1005,13 +1009,14 @@ splitResult f sc@(SClause tel ps _ _ target) = do
                          }
             return (SplitCon (unArg proj), sc')
       _ -> failure $ CosplitNoRecordType tel $ unArg t
-  where
-  -- A record type is strong if it has all the projections.
-  -- This is the case if --irrelevant-projections or no field is irrelevant.
-  -- TODO: what about shape irrelevance?
-  strongRecord :: [Arg QName] -> TCM Bool
-  strongRecord fs = (optIrrelevantProjections <$> pragmaOptions) `or2M`
-    (return $ not $ any isIrrelevant fs)
+  -- Andreas, 2018-06-09, issue #2170: splitting with irrelevant fields is always fine!
+  -- where
+  -- -- A record type is strong if it has all the projections.
+  -- -- This is the case if --irrelevant-projections or no field is irrelevant.
+  -- -- TODO: what about shape irrelevance?
+  -- strongRecord :: [Arg QName] -> TCM Bool
+  -- strongRecord fs = (optIrrelevantProjections <$> pragmaOptions) `or2M`
+  --   (return $ not $ any isIrrelevant fs)
 
 
 -- * Boring instances